+++ blocks/init/src/blocks/custom/button/button-transforms.js
@@ -1,7 +1,7 @@
 import { createBlock } from '@wordpress/blocks';
-import manifest from './../../manifest.json';
+import manifest from '../../manifest.json';
 import manifestButton from './manifest.json';
-import manifestLink from './../link/manifest.json';
+import manifestLink from '../link/manifest.json';
 
 export const Transforms = {
   from: [
