+++ blocks/init/src/blocks/custom/column/docs/story.js
@@ -5,7 +5,6 @@ import React from 'react';
 import manifest from './../manifest.json';
 import globalManifest from './../../../manifest.json';
 import readme from './readme.md';
-import { block as item } from '../../card/docs/story';
 
 export default {
   title: `Blocks|${manifest.title}`,
@@ -15,5 +14,5 @@ export default {
 };
 
 export const block = () => (
-  <Gutenberg props={blockDetails(manifest, globalManifest, item(), 1)} />
+  <Gutenberg props={blockDetails(manifest, globalManifest)} />
 );
