+++ blocks/init/src/blocks/custom/heading/heading-transforms.js
@@ -1,7 +1,7 @@
 import { createBlock } from '@wordpress/blocks';
-import manifest from './../../manifest.json';
+import manifest from '../../manifest.json';
 import manifestHeading from './manifest.json';
-import manifestParagraph from './../paragraph/manifest.json';
+import manifestParagraph from '../paragraph/manifest.json';
 
 export const Transforms = {
   from: [
