+++ blocks/init/src/blocks/custom/image/image-transforms.js
@@ -1,7 +1,7 @@
 import { createBlock } from '@wordpress/blocks';
-import manifest from './../../manifest.json';
+import manifest from '../../manifest.json';
 import manifestImage from './manifest.json';
-import manifestCarouselImage from './../carousel-image/manifest.json';
+import manifestCarouselImage from '../carousel-image/manifest.json';
 
 export const Transforms = {
   from: [
