+++ blocks/init/src/blocks/custom/link/link-transforms.js
@@ -1,7 +1,7 @@
 import { createBlock } from '@wordpress/blocks';
-import manifest from './../../manifest.json';
+import manifest from '../../manifest.json';
 import manifestLink from './manifest.json';
-import manifestButton from './../button/manifest.json';
+import manifestButton from '../button/manifest.json';
 
 export const Transforms = {
   from: [
