+++ blocks/init/src/blocks/custom/paragraph/paragraph-transforms.js
@@ -1,7 +1,7 @@
 import { createBlock } from '@wordpress/blocks';
-import manifest from './../../manifest.json';
+import manifest from '../../manifest.json';
 import manifestParagraph from './manifest.json';
-import manifestHeading from './../heading/manifest.json';
+import manifestHeading from '../heading/manifest.json';
 
 export const Transforms = {
   from: [
